# swagger-docs.json
{"openapi":"3.0.0","info":{"title":"Boiterplate Api","description":"Boiterplate Api","version":"0.0.1","contact":{}},"tags":[],"servers":[{"url":"http://localhost:3333/"}],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDTO":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"LoggedUserResponse":{"type":"object","properties":{"jwtToken":{"type":"string"},"user":{"type":"object"}},"required":["jwtToken","user"]},"RegisterDTO":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]}}},"paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/auth/login":{"post":{"operationId":"AuthenticationsController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"200":{"description":"Login","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoggedUserResponse"}}}}},"tags":["auth"]}},"/auth/admins":{"post":{"operationId":"AuthenticationsController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDTO"}}}},"responses":{"200":{"description":"Register Admin","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoggedUserResponse"}}}}},"tags":["auth"]}},"/auth/clients":{"post":{"operationId":"AuthenticationsController_registerClient","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDTO"}}}},"responses":{"200":{"description":"Register Client","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoggedUserResponse"}}}}},"tags":["auth"]}},"/auth/me":{"get":{"operationId":"AuthenticationsController_me","parameters":[],"responses":{"200":{"description":"Return the logged user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoggedUserResponse"}}}}},"tags":["auth"],"security":[{"bearer":[]}]}}}}